Enhanced parse by using doubles over floats for decimal datatypes, model generation script in package.json that uses a ruby script, a recreate:dev_test script that generates and migrates dev and test databases in one shot and updated migrations and models to reflect changes mentioned above in the ruby script (#34)

+++ migrations/20180404193957-create-experiment.js
@@ -75,7 +75,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       dilution_solvent_percent_max: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DOUBLE
       },
       assay_component_name: {
         type: Sequelize.STRING
